feat: 新增BasePage类，主窗口增加添加页面的接口。
